+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+#2015-11-09
+ - Allow complex expressions in route parameters
+
 #2014-04-18
  - Rewrite documentation